spread spectrum clocking -- confirmed working via spectrum analyzer
Chain an MMCME to a PLL to get spread spectrum clocking to work.

why? Because SS clocking only works with some really strange
multiples that don't allow you to generate the clocks you actually need.
So we use an MMCME2 to whiten up the incoming clock and then pass
it on to the PLLE2 block to multiply up for consumption by the rest
of the FPGA.

This diminishes, somewhat, the efficacy of the spread spectrum
clocking because the PLLE2 will eat up some of the frequency wander.
However, when viewed on a spectrum analyzer it seems to effectively
spread the energy out a bit, probably by enough to make a difference
for EMI/EMC certification.