deps: site: undo the brokenness that is python eval()
The Python site.py module calls exec(), which is the Python equivalent
of eval().  This gets run on .pth files, which contain lists of eggs
to append to the system path.

As a result of this eval(), the module ordering gets completely broken.

Undo this by cleaning up after the eval(), and inserting anything it
added /after/ the current search order.

Signed-off-by: Sean Cross <sean@xobs.io>